requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymupdf==1.23.5
qdrant-client==1.8.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pillow==10.3.0
sentence-transformers==2.2.2
click==8.1.7